Change package build script to build from a tarball.
Instead of building from within the source directory, it now expects a
source tarball in the current directory to be specified, and to have a
debian directory available at the current level (which may have a
distribution specific suffix).

For my workflow I prefer to build from a distribution tarball as it
makes it much harder to accidentally leave important files out of the
distribution.